flask_advanced_project/requirements.txt
# Flask and Extensions
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-pymongo==2.3.0
flask-bcrypt==1.0.1
flask-limiter==3.5.0

# MongoDB
pymongo==4.6.0

# Environment and Configuration
python-dotenv==1.0.0
dynaconf==3.2.3

# Validation and Serialization
marshmallow==3.20.1

# Utilities
pytz==2023.3
Werkzeug==2.3.7
requests==2.31.0
tenacity==8.2.3
python-dateutil==2.8.2

# Security
bcrypt==4.0.1
passlib==1.7.4
python-jose==3.3.0
crypto==1.4.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
factory-boy==3.3.0

# Development Tools
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0

# Documentation
Sphinx==7.2.6
flasgger==0.9.7.1
